fix(@DataTable): sort spec
